搞定公司、租户等的维护

+++ service/service-pcm/src/main/java/org/github/foxnic/web/pcm/controller/CatalogController.java
@@ -1,47 +1,44 @@
 package org.github.foxnic.web.pcm.controller;
 
- 
-import java.util.List;
 
-import org.springframework.web.bind.annotation.RestController;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.web.bind.annotation.PostMapping;
-
-import org.github.foxnic.web.framework.web.SuperController;
-import org.github.foxnic.web.framework.sentinel.SentinelExceptionUtil;
-import org.springframework.web.bind.annotation.RequestMapping;
-import javax.servlet.http.HttpServletResponse;
-import org.springframework.web.multipart.MultipartFile;
-import org.springframework.web.multipart.MultipartHttpServletRequest;
 import com.alibaba.csp.sentinel.annotation.SentinelResource;
-
-
-import org.github.foxnic.web.proxy.pcm.CatalogServiceProxy;
-import org.github.foxnic.web.domain.pcm.meta.CatalogVOMeta;
-import org.github.foxnic.web.domain.pcm.Catalog;
-import org.github.foxnic.web.domain.pcm.CatalogVO;
+import com.github.foxnic.api.error.ErrorDesc;
 import com.github.foxnic.api.transter.Result;
+import com.github.foxnic.api.validate.annotations.NotNull;
+import com.github.foxnic.commons.io.StreamUtil;
+import com.github.foxnic.commons.lang.StringUtil;
+import com.github.foxnic.dao.data.PagedList;
 import com.github.foxnic.dao.data.SaveMode;
 import com.github.foxnic.dao.excel.ExcelWriter;
-import com.github.foxnic.springboot.web.DownloadUtil;
-import com.github.foxnic.dao.data.PagedList;
-import java.util.Date;
-import java.sql.Timestamp;
-import com.github.foxnic.api.error.ErrorDesc;
-import com.github.foxnic.commons.io.StreamUtil;
-import java.util.Map;
 import com.github.foxnic.dao.excel.ValidateResult;
-import java.io.InputStream;
-import org.github.foxnic.web.domain.pcm.meta.CatalogMeta;
-import io.swagger.annotations.Api;
-import com.github.xiaoymin.knife4j.annotations.ApiSort;
-import io.swagger.annotations.ApiOperation;
-import io.swagger.annotations.ApiImplicitParams;
-import io.swagger.annotations.ApiImplicitParam;
+import com.github.foxnic.springboot.web.DownloadUtil;
 import com.github.xiaoymin.knife4j.annotations.ApiOperationSupport;
-import com.alibaba.csp.sentinel.annotation.SentinelResource;
+import com.github.xiaoymin.knife4j.annotations.ApiSort;
+import io.swagger.annotations.Api;
+import io.swagger.annotations.ApiImplicitParam;
+import io.swagger.annotations.ApiImplicitParams;
+import io.swagger.annotations.ApiOperation;
+import org.github.foxnic.web.domain.oauth.MenuVO;
+import org.github.foxnic.web.domain.oauth.meta.MenuVOMeta;
+import org.github.foxnic.web.domain.pcm.Catalog;
+import org.github.foxnic.web.domain.pcm.CatalogVO;
+import org.github.foxnic.web.domain.pcm.meta.CatalogVOMeta;
+import org.github.foxnic.web.framework.sentinel.SentinelExceptionUtil;
+import org.github.foxnic.web.framework.web.SuperController;
+import org.github.foxnic.web.misc.ztree.ZTreeNode;
 import org.github.foxnic.web.pcm.service.ICatalogService;
-import com.github.foxnic.api.validate.annotations.NotNull;
+import org.github.foxnic.web.proxy.pcm.CatalogServiceProxy;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.multipart.MultipartFile;
+import org.springframework.web.multipart.MultipartHttpServletRequest;
+
+import javax.servlet.http.HttpServletResponse;
+import java.io.InputStream;
+import java.util.List;
+import java.util.Map;
 
 /**
  * <p>
@@ -49,6 +46,7 @@ import com.github.foxnic.api.validate.annotations.NotNull;
  * </p>
  * @author 李方捷 , leefangjie@qq.com
  * @since 2021-08-28 15:45:05
+ * @version 20210830
 */
 
 @Api(tags = "数据存储")
@@ -77,7 +75,15 @@ public class CatalogController extends SuperController {
 	@SentinelResource(value = CatalogServiceProxy.INSERT , blockHandlerClass = { SentinelExceptionUtil.class } , blockHandler = SentinelExceptionUtil.HANDLER )
 	@PostMapping(CatalogServiceProxy.INSERT)
 	public Result insert(CatalogVO catalogVO) {
+
+		catalogVO.setSort(999999);
+		if(StringUtil.isBlank(catalogVO.getParentId())) {
+			catalogVO.setParentId(ICatalogService.ROOT_ID);
+		}
 		Result result=catalogService.insert(catalogVO);
+		if(result.success()) {
+			result.data(catalogVO);
+		}
 		return result;
 	}
 
@@ -93,8 +99,14 @@ public class CatalogController extends SuperController {
 	@NotNull(name = CatalogVOMeta.ID)
 	@SentinelResource(value = CatalogServiceProxy.DELETE , blockHandlerClass = { SentinelExceptionUtil.class } , blockHandler = SentinelExceptionUtil.HANDLER )
 	@PostMapping(CatalogServiceProxy.DELETE)
-	public Result deleteById(Integer id) {
-		Result result=catalogService.deleteByIdLogical(id);
+	public Result deleteById(String id) {
+		Result result = null;
+		List children=catalogService.queryChildNodes(id);
+		if(!children.isEmpty()) {
+			result.success(false).message("请先删除下级节点");
+			return result;
+		}
+		result=catalogService.deleteByIdLogical(id);
 		return result;
 	}
 	
@@ -171,9 +183,14 @@ public class CatalogController extends SuperController {
 	@NotNull(name = CatalogVOMeta.ID)
 	@SentinelResource(value = CatalogServiceProxy.GET_BY_ID , blockHandlerClass = { SentinelExceptionUtil.class } , blockHandler = SentinelExceptionUtil.HANDLER )
 	@PostMapping(CatalogServiceProxy.GET_BY_ID)
-	public Result<Catalog> getById(Integer id) {
+	public Result<Catalog> getById(String id) {
 		Result<Catalog> result=new Result<>();
 		Catalog catalog=catalogService.getById(id);
+		List children=catalogService.queryChildNodes(id);
+		if(!children.isEmpty()) {
+			result.success(false).message("请先删除下级节点");
+			return result;
+		}
 		result.success(true).data(catalog);
 		return result;
 	}
@@ -191,7 +208,7 @@ public class CatalogController extends SuperController {
 		@NotNull(name = CatalogVOMeta.IDS)
 		@SentinelResource(value = CatalogServiceProxy.GET_BY_IDS , blockHandlerClass = { SentinelExceptionUtil.class } , blockHandler = SentinelExceptionUtil.HANDLER )
 	@PostMapping(CatalogServiceProxy.GET_BY_IDS)
-	public Result<List<Catalog>> getByIds(List<Integer> ids) {
+	public Result<List<Catalog>> getByIds(List<String> ids) {
 		Result<List<Catalog>> result=new Result<>();
 		List<Catalog> list=catalogService.getByIds(ids);
 		result.success(true).data(list);
@@ -217,6 +234,79 @@ public class CatalogController extends SuperController {
 		Result<List<Catalog>> result=new Result<>();
 		List<Catalog> list=catalogService.queryList(sample);
 		result.success(true).data(list);
+		return result;
+	}
+
+
+	/**
+	 * 查询数据存储
+	 */
+	@ApiOperation(value = "查询数据存储")
+	@ApiImplicitParams({
+			@ApiImplicitParam(name = CatalogVOMeta.ID , value = "主键" , required = true , dataTypeClass=Integer.class),
+			@ApiImplicitParam(name = CatalogVOMeta.NAME , value = "名称" , required = false , dataTypeClass=String.class),
+			@ApiImplicitParam(name = CatalogVOMeta.PARENT_ID , value = "上级ID" , required = false , dataTypeClass=Integer.class),
+			@ApiImplicitParam(name = CatalogVOMeta.DATA_TABLE , value = "存储表" , required = false , dataTypeClass=String.class),
+			@ApiImplicitParam(name = CatalogVOMeta.TENANT_ID , value = "租户ID" , required = true , dataTypeClass=Integer.class),
+	})
+	@ApiOperationSupport(order=5 ,  ignoreParameters = { CatalogVOMeta.PAGE_INDEX , CatalogVOMeta.PAGE_SIZE } )
+	@SentinelResource(value = CatalogServiceProxy.QUERY_NODES , blockHandlerClass = { SentinelExceptionUtil.class } , blockHandler = SentinelExceptionUtil.HANDLER )
+	@PostMapping(CatalogServiceProxy.QUERY_NODES)
+	public Result<List<ZTreeNode>> queryNodes(MenuVO sample) {
+
+		Result<List<ZTreeNode>> result=new Result<>();
+		List<ZTreeNode> list=null;
+		if(sample.getParentId()==null) {
+			list=catalogService.queryRootNotes();
+		} else {
+			list=catalogService.queryChildNodes(sample.getParentId());
+		}
+		//加载全部子孙节点
+		if(sample.getIsLoadAllDescendants()!=null && sample.getIsLoadAllDescendants()==1) {
+
+			list=catalogService.buildingHierarchicalRelationships(list);
+
+		}
+		result.success(true).data(list);
+		return result;
+
+	}
+
+
+	/**
+	 * 变更分类层级关系
+	 */
+	@ApiOperation(value = "变更分类层级关系")
+	@ApiImplicitParams({
+			@ApiImplicitParam(name = "ids" , value = "ID" , required = true , dataTypeClass=String.class , example = "451739184575545344"),
+			@ApiImplicitParam(name = MenuVOMeta.PARENT_ID , value = "新的上级节点ID" , required = true , dataTypeClass=String.class , example = "451739184575545344")
+	})
+	@ApiOperationSupport(order=2)
+	@NotNull(name = "ids")
+	@SentinelResource(value = CatalogServiceProxy.SAVE_HIERARCHY)
+	@PostMapping(CatalogServiceProxy.SAVE_HIERARCHY)
+	public Result<Catalog> changeParent(List<String> ids, String parentId) {
+		Result<Catalog> result=new Result<>();
+		Boolean suc=catalogService.saveHierarchy(ids,parentId);
+		result.success(suc);
+		return result;
+	}
+
+
+	/**
+	 * 变更分类层级关系
+	 */
+	@ApiOperation(value = "变更分类层级关系")
+	@ApiImplicitParams({
+			@ApiImplicitParam(name = "keyword" , value = "keyword" , required = true , dataTypeClass=String.class , example = "橡胶")
+	})
+	@ApiOperationSupport(order=2)
+	@SentinelResource(value = CatalogServiceProxy.SEARCH)
+	@PostMapping(CatalogServiceProxy.SEARCH)
+	public Result<List<String>> search(String keyword) {
+		Result<List<String>> result=new Result<>();
+		List<String> hierarchyList=catalogService.search(keyword);
+		result.data(hierarchyList);
 		return result;
 	}
